# api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Digitalisierungsfabrik API",
    "description": "AI-guided business process elicitation system. Backend for the Digitalisierungsfabrik prototype.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "description": "Alle Projekte auflisten (FR-G-02).",
        "operationId": "list_projects_api_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "description": "Neues Projekt anlegen (FR-G-01).\n\nPer SDD 6.1.0: Neue Projekte starten im Moderator-Modus.\nDie Begrüßung erfolgt über den WebSocket bei Verbindungsaufbau (FR-D-11),\nNICHT synchron hier — damit der Moderator im Modus bleibt und der\nNutzer mit ihm interagieren kann, bevor die Exploration beginnt.",
        "operationId": "create_project_api_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projekt_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "description": "Projektdetails laden (FR-E-02).",
        "operationId": "get_project_api_projects__projekt_id__get",
        "parameters": [
          {
            "name": "projekt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projekt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Project",
        "description": "Einzelnes Projekt löschen (FR-E-06: isoliert).",
        "operationId": "delete_project_api_projects__projekt_id__delete",
        "parameters": [
          {
            "name": "projekt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projekt Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projekt_id}/artifacts": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Artifacts",
        "description": "Alle drei Artefakte eines Projekts laden (FR-B-06).",
        "operationId": "get_artifacts_api_projects__projekt_id__artifacts_get",
        "parameters": [
          {
            "name": "projekt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projekt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactsResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projekt_id}/download": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Download Project",
        "description": "Alle drei Artefakte als JSON herunterladen (FR-B-07).",
        "operationId": "download_project_api_projects__projekt_id__download_get",
        "parameters": [
          {
            "name": "projekt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projekt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/batch": {
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Projects Batch",
        "description": "Mehrere Projekte gleichzeitig löschen.",
        "operationId": "delete_projects_batch_api_projects_batch_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectDeleteBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDeleteBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projekt_id}/complete": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Complete Project",
        "description": "Projekt als abgeschlossen markieren (FR-G-04).",
        "operationId": "complete_project_api_projects__projekt_id__complete_post",
        "parameters": [
          {
            "name": "projekt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projekt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCompleteResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projekt_id}/artifacts/{typ}/versions": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Artifact Versions",
        "description": "Versionshistorie eines Artefakts (FR-B-10).",
        "operationId": "list_artifact_versions_api_projects__projekt_id__artifacts__typ__versions_get",
        "parameters": [
          {
            "name": "projekt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projekt Id"
            }
          },
          {
            "name": "typ",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Typ"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactVersionsResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/projects/{projekt_id}/artifacts/{typ}/restore": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Restore Artifact Version",
        "description": "Version wiederherstellen — erzeugt neue Version (FR-B-10).",
        "operationId": "restore_artifact_version_api_projects__projekt_id__artifacts__typ__restore_post",
        "parameters": [
          {
            "name": "projekt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projekt Id"
            }
          },
          {
            "name": "typ",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Typ"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactRestoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactRestoreResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/projects/{projekt_id}/import": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Import Artifact",
        "description": "Artefakt importieren und validieren (FR-B-07, FR-C-04).",
        "operationId": "import_artifact_api_projects__projekt_id__import_post",
        "parameters": [
          {
            "name": "projekt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projekt Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactRestoreResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/projects/{projekt_id}/debug/advance-phase": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Debug Advance Phase",
        "operationId": "debug_advance_phase_api_projects__projekt_id__debug_advance_phase_post",
        "parameters": [
          {
            "name": "projekt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projekt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvancePhaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projekt_id}/validation": {
      "get": {
        "tags": [
          "projects",
          "projects"
        ],
        "summary": "Get Validation Report",
        "description": "Aktuellen Validierungsbericht abrufen (SDD 6.6.4, FR-C-08).",
        "operationId": "get_validation_report_api_projects__projekt_id__validation_get",
        "parameters": [
          {
            "name": "projekt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projekt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReportResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projekt_id}/export": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Export Project",
        "description": "Export all three artifacts as JSON + Markdown (Story 11-02, FR-B-07).",
        "operationId": "export_project_api_projects__projekt_id__export_get",
        "parameters": [
          {
            "name": "projekt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projekt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProjectCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Name"
          },
          "beschreibung": {
            "type": "string",
            "title": "Beschreibung",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectCreateRequest",
        "description": "Request body for POST /api/projects (FR-G-01)."
      },
...
      "ProjectListResponse": {
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            },
            "type": "array",
            "title": "Projects"
          }
        },
        "type": "object",
        "required": [
          "projects"
        ],
        "title": "ProjectListResponse",
        "description": "Response for GET /api/projects (FR-G-02)."
      },
      "ProjectResponse": {
        "properties": {
          "projekt_id": {
            "type": "string",
            "title": "Projekt Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "beschreibung": {
            "type": "string",
            "title": "Beschreibung"
          },
          "erstellt_am": {
            "type": "string",
            "format": "date-time",
            "title": "Erstellt Am"
          },
          "zuletzt_geaendert": {
            "type": "string",
            "format": "date-time",
            "title": "Zuletzt Geaendert"
          },
          "aktive_phase": {
            "$ref": "#/components/schemas/Projektphase"
          },
          "aktiver_modus": {
            "type": "string",
            "title": "Aktiver Modus"
          },
          "projektstatus": {
            "$ref": "#/components/schemas/Projektstatus"
          }
        },
        "type": "object",
        "required": [
          "projekt_id",
          "name",
          "beschreibung",
          "erstellt_am",
          "zuletzt_geaendert",
          "aktive_phase",
          "aktiver_modus",
          "projektstatus"
        ],
        "title": "ProjectResponse",
        "description": "Project metadata — all SDD 7.2.1 fields."
      },
      "Projektphase": {
        "type": "string",
        "enum": [
          "exploration",
          "strukturierung",
          "spezifikation",
          "validierung",
          "abgeschlossen"
        ],
        "title": "Projektphase",
        "description": "Aktive Phase des Projekts."
      },
...
    }
  }
}